+++ src/Game/Systems/SpawnerSystem.cpp
@@ -30,7 +30,7 @@ EntityWrapper SpawnerSystem::Spawn(EntityWrapper spawner, EntityWrapper parent /
         if (spawnPoints.size() > 1) {
             static std::random_device randomDevice;
             static std::mt19937 randomGenerator(randomDevice());
-            std::uniform_int_distribution<> distribution(0, std::distance(spawnPoints.begin(), spawnPoints.end()) - 1);
+            std::uniform_int_distribution<> distribution(0, static_cast<int>(std::distance(spawnPoints.begin(), spawnPoints.end())) - 1);
             auto randomSpawnPointIt = spawnPoints.begin();
             std::advance(randomSpawnPointIt, distribution(randomGenerator));
             spawnPoint = *randomSpawnPointIt;
